Refactor: Create dedicated exceptions module and improve cmake_format.py
